internaljobmarket/requirements.txt

flask
Flask-SQLAlchemy
Flask-WTF